+++ backend/canisters/notifications_index/api/src/queries/fcm_token_exists.rs
@@ -1,11 +1,12 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
+use types::FcmToken;
 
 #[ts_export(notifications_index, fcm_token_exists)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub fcm_token: String,
+    pub fcm_token: FcmToken,
 }
 
 pub type Response = bool;
