+++ backend/libraries/types/src/lib.rs
@@ -39,6 +39,7 @@ mod event_wrapper;
 mod events;
 mod exchange_id;
 mod exchanges;
+mod fcm_token;
 mod field_invalid;
 mod file;
 mod file_status;
@@ -116,6 +117,7 @@ pub use event_wrapper::*;
 pub use events::*;
 pub use exchange_id::*;
 pub use exchanges::*;
+pub use fcm_token::*;
 pub use field_invalid::*;
 pub use file::*;
 pub use file_status::*;
